requirements.txt
# Tencent Cloud SDK for translation
PyYAML>=6.0
toml>=0.10.2
tencentcloud-sdk-python>=3.0.0
requests>=2.28.0